+++ CHANGES.md
@@ -7,6 +7,7 @@
 has required fields, reference files exist, FASTQ files match libraries.tsv,
 bowtie2 index is not required if aligner strobealign is used.
 * More verbose command line messages when an error occurs.
+* More verbose docstring for `minute run` command.
 
 ## v0.9.1